Moved filter textfield to be above header status bar
