+++ src/main/java/address/controller/RootLayoutController.java
@@ -1,13 +1,19 @@
 package address.controller;
 
 import address.MainApp;
+import address.events.parser.FilterCommittedEvent;
 import address.keybindings.KeyBindingsManager;
 import address.model.ModelManager;
+import address.parser.ParseException;
+import address.parser.Parser;
+import address.parser.expr.Expr;
+import address.parser.expr.PredExpr;
 import address.util.AppLogger;
 import address.util.LoggerManager;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.MenuItem;
+import javafx.scene.control.TextField;
 
 /**
  * The controller for the root layout. The root layout provides the basic
@@ -21,11 +27,17 @@ public class RootLayoutController extends UiController {
     private ModelManager modelManager;
     private MainApp mainApp;
 
+    private Parser parser;
+
     @FXML
     private MenuItem helpMenuItem;
 
+    @FXML
+    private TextField filterField;
+
     public RootLayoutController() {
         super();
+        parser = new Parser();
     }
 
     public void setConnections(MainApp mainApp, MainController mainController, ModelManager modelManager) {
@@ -36,6 +48,21 @@ public class RootLayoutController extends UiController {
 
     public void setAccelerators() {
         helpMenuItem.setAccelerator(KeyBindingsManager.getAcceleratorKeyCombo("HELP_PAGE_ACCELERATOR").get());
+    }
+
+    @FXML
+    private void handleFilterChanged() {
+        Expr filterExpression;
+        try {
+            filterExpression = parser.parse(filterField.getText());
+            if (filterField.getStyleClass().contains("error")) filterField.getStyleClass().remove("error");
+        } catch (ParseException e) {
+            logger.debug("Invalid filter found: {}", e);
+            filterExpression = PredExpr.TRUE;
+            if (!filterField.getStyleClass().contains("error")) filterField.getStyleClass().add("error");
+        }
+
+        raise(new FilterCommittedEvent(filterExpression));
     }
 
     @FXML
