+++ src/main/resources/view/PersonListPanel.fxml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.*?>
 <?import java.net.URL?>
 <VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="address.controller.PersonListPanelController" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
@@ -10,7 +9,6 @@
         <URL value="@Extensions.css" />
     </stylesheets>
     <children>
-        <TextField fx:id="filterField" onAction="#handleFilterChanged" />
         <ListView fx:id="personListView" VBox.vgrow="ALWAYS" />
     </children>
 </VBox>
