+++ src/main/resources/view/RootLayout.fxml
@@ -2,7 +2,12 @@
 
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
+<?import java.net.URL?>
 <VBox maxHeight="Infinity" maxWidth="Infinity" minHeight="-Infinity" minWidth="-Infinity" stylesheets="@DarkTheme.css" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="address.controller.RootLayoutController">
+   <stylesheets>
+      <URL value="@DarkTheme.css" />
+      <URL value="@Extensions.css" />
+   </stylesheets>
    <children>
       <MenuBar VBox.vgrow="NEVER">
          <menus>
@@ -24,6 +29,7 @@
             </Menu>
          </menus>
       </MenuBar>
+      <TextField fx:id="filterField" onAction="#handleFilterChanged" />
       <AnchorPane fx:id="headerStatusbarPlaceholder" styleClass="anchor-pane-with-border">
       </AnchorPane>
       <SplitPane id="splitPane" fx:id="splitPane" dividerPositions="0.4" VBox.vgrow="ALWAYS">
